[feat] 설정 정보 변경

+++ aspot-auction/src/main/java/com/spot/auction/aspot_auction/dto/AuctionItemRequest.java
@@ -2,7 +2,6 @@ package com.spot.auction.aspot_auction.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -10,7 +9,7 @@ public class AuctionItemRequest {
     private String title;
     private String description;
     private String imageUrl;
-    private BigDecimal startPrice;
+    private Integer startPrice;
     private String category;
     private LocalDate auctionDate;
     private String status;
