+++ aspot-auction/src/main/java/com/spot/auction/aspot_auction/model/AuctionItem.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -19,7 +18,7 @@ public class AuctionItem {
     private String title;
     private String description;
     private String imageUrl;
-    private BigDecimal startPrice;
+    private Integer startPrice;
     private String category;
     private LocalDate auctionDate;
     private String status;
